Add support for temporal dispatching, add support for initiation function, correct a bunch of errors in the interpretation of the v7em spec

+++ symex/src/executor/extension/ieee754.rs
@@ -47,7 +47,7 @@ where
             }
             OperandStorage::Address(address) => {
                 let address = extract!(Ok(self.get_operand_value(&address, logger)));
-                let read = match self.state.reader().read_memory(address.clone(), operand.ty.size() as usize) {
+                let read = match self.state.reader().read_memory(address.clone(), operand.ty.size()) {
                     crate::executor::hooks::ResultOrHook::Hook(hook) => match hook(&mut self.state, address) {
                         Ok(val) => val,
                         Err(e) => return ResultOrTerminate::Result(Err(e)),
